grouped ui checks in a module
